requirements.txt
miniworld==2.1.0
gymnasium>=0.29.0,<1.1.0
numpy>=1.24.0,<2.0.0
stable-baselines3>=2.0.0
torch>=2.0.0
pygame>=2.5.0
Pillow>=10.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0
tqdm>=4.65.0